Scraper/requirements.txt
requests
redis
psycopg2-binary  # Use pre-compiled binary